Eliminate vue-chartjs
It gets in the way between the app and chart.js (easier to just read the
chart.js docs and not have to translate).  Also, it was causing problems
when I tried to use production (minified) vue.js, because it was giving
me warnings about using development vue.js until I removed it.