Read all memory regions before updating
Changed the websock message handler to finish reading all the memory
regions before doing an update rather than updating them as they come.

This should hopefully fix a bug reported by strotlog in which two or
more memory regions come in the same message (this never happens with
retroarch -- which is why I did not see the same behavior -- but is
likely when using sd2snes).